added delete feature for team events

+++ src/buttonHandler/buttonFunctions/teamEvent/deleteTeam.ts
@@ -10,7 +10,10 @@ import {
 
 import { ButtonFunction } from '../../ButtonStructure';
 import infoMessageEmbed from '../../../globalUtils/infoMessageEmbed';
-import { deleteTeam as deleteTeamSupabase } from '../../../supabase/supabaseFunctions/teams';
+import {
+	deleteTeam as deleteTeamSupabase,
+	checkTeamExists,
+} from '../../../supabase/supabaseFunctions/teams';
 
 const deleteTeam: ButtonFunction = {
 	customId: 'deleteTeam',
@@ -33,12 +36,12 @@ const deleteTeam: ButtonFunction = {
 				});
 			}
 
+			const teamId: string | any =
+				interaction.message.embeds[0].footer?.text.split(': ')[2];
+
 			const fetchedMessage = await interaction.channel?.messages.fetch(
 				interaction.message.id,
 			);
-
-			const teamId: string | any =
-				interaction.message.embeds[0].footer?.text.split(': ')[2];
 
 			if (fetchedMessage) {
 				const confirmationButtons = new MessageActionRow().addComponents(
@@ -80,7 +83,8 @@ const deleteTeam: ButtonFunction = {
 
 						await interaction.deleteReply();
 
-						deleteTeamSupabase({ teamId: parseInt(teamId) });
+						if (await checkTeamExists({ teamId: teamId }))
+							deleteTeamSupabase({ teamId: parseInt(teamId) });
 
 						await i.reply({
 							embeds: [
